move tag management functions to Purebred.Storage.Tags
