+++ test/TestMail.hs
@@ -35,9 +35,8 @@ import Test.QuickCheck.Instances ()
 import Notmuch (mkTag, tagMaxLen)
 
 import Purebred.Types
-import Purebred.Storage.Notmuch (addTags, removeTags, tagItem)
 import Purebred.Storage.Mail (findMatchingWords)
-import Purebred.Storage.Tags (TagOp(..))
+import Purebred.Storage.Tags (TagOp(..), addTags, removeTags, tagItem)
 
 mailTests ::
   TestTree
